+++ Changelog.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased: 5.1.0-rc.17]
+ - FIO-9385 Preserve non-default widget settings after evaluating field logic
+ 
 ## 5.1.0-rc.16
 ### Changed
  - FIO-9515 Fixed Edit grid conditional values not displayed in Edit sub…